requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
python-telegram-bot-calendar==1.0.5
python-telegram-bot-pagination==0.0.2
peewee==3.15.4
loguru==0.6.0